+++ tools/byte_converter/src/hooks/use-masked-value.ts
@@ -1,8 +1,16 @@
 import { useCallback, useMemo, useState } from "preact/hooks";
 import { Unit } from "../types";
 
-const HighByteMask = 65280; // %1111111100000000
-const LowByteMask = 255; // %0000000011111111
+const Mask = {
+  [Unit.Byte]: {
+    prev: 65280, // %1111111100000000
+    next: 255, // %0000000011111111
+  },
+  [Unit.Word]: {
+    prev: 0, // %0000000000000000
+    next: 65535, // %1111111111111111
+  },
+};
 
 export default function useMaskedValue(
   initialValue: number,
@@ -10,17 +18,13 @@ export default function useMaskedValue(
 ): [number, (nextValue: number) => void] {
   const [value, setValue] = useState(initialValue);
 
-  const maskedValue = useMemo(
-    () => (unit === Unit.Byte ? value & LowByteMask : value),
-    [unit, value]
-  );
+  const maskedValue = useMemo(() => value & Mask[unit].next, [unit, value]);
 
   const setMaskedValue = useCallback(
     (nextValue: number) => {
       setValue((prevValue) => {
-        return unit === Unit.Byte
-          ? (prevValue & HighByteMask) | (nextValue & LowByteMask)
-          : nextValue;
+        const mask = Mask[unit];
+        return (prevValue & mask.prev) | (nextValue & mask.next);
       });
     },
     [unit]